fix: address review fixes and use same file structure like map index
