+++ lib/segment/src/index/field_index/geo_index/mmap_geo_index/lifecycle.rs
@@ -4,7 +4,7 @@ use std::path::{Path, PathBuf};
 
 use ahash::AHashSet;
 use common::binary_search::binary_search_by;
-use common::bitvec::{BitSlice, BitSliceExt, BitVec};
+use common::bitvec::{BitSlice, BitSliceExt};
 use common::counter::conditioned_counter::ConditionedCounter;
 use common::counter::hardware_counter::HardwareCounterCell;
 use common::fs::{atomic_save_json, clear_disk_cache, read_json};
@@ -16,114 +16,17 @@ use common::types::PointOffsetType;
 use common::universal_io::{MmapFile, OpenOptions, ReadRange, TypedStorage, UniversalRead};
 use fs_err as fs;
 use memmap2::MmapMut;
-use serde::{Deserialize, Serialize};
 
-use super::mutable_geo_index::InMemoryGeoMapIndex;
-use super::read_ops::GeoMapIndexRead;
+use super::super::mutable_geo_index::InMemoryGeoMapIndex;
+use super::{
+    COUNTS_PER_HASH, Counts, DELETED_PATH, POINTS_MAP, POINTS_MAP_IDS, PointKeyValue, STATS_PATH,
+    Storage, StoredGeoMapIndex, StoredGeoMapIndexStat,
+};
 use crate::common::Flusher;
 use crate::common::operation_error::{OperationError, OperationResult};
 use crate::index::field_index::geo_hash::{GeoHash, GeoHashRaw};
 use crate::index::field_index::stored_point_to_values::StoredPointToValues;
-use crate::index::payload_config::StorageType;
 use crate::types::GeoPoint;
-
-const DELETED_PATH: &str = "deleted.bin";
-const COUNTS_PER_HASH: &str = "counts_per_hash.bin";
-const POINTS_MAP: &str = "points_map.bin";
-const POINTS_MAP_IDS: &str = "points_map_ids.bin";
-const STATS_PATH: &str = "mmap_field_index_stats.json";
-
-#[repr(C)]
-#[derive(Copy, Clone, Debug, bytemuck::Pod, bytemuck::Zeroable)]
-pub(super) struct Counts {
-    pub hash: GeoHashRaw,
-    pub points: u32,
-    pub values: u32,
-}
-
-#[repr(C)]
-#[derive(Copy, Clone, Debug, bytemuck::Pod, bytemuck::Zeroable)]
-pub(super) struct PointKeyValue {
-    pub hash: GeoHashRaw,
-    pub ids_start: u32,
-    pub ids_end: u32,
-}
-
-///
-///   points_map
-///  ┌─────────────────────────────────────────┐
-///  │ (ABC, 10, 20)|(ABD, 20, 40)             │
-///  └────────┬──┬──────────┬───┬──────────────┘
-///           │  │          │   │
-///    ┌──────┘  └────────┐ │   └───────────────────┐
-///    │                  │ └───┐                   │
-///    │                  │     │                   │
-///  ┌─▼──────────────────▼─────▼───────────────────▼──────────┐
-///  │ 1, 8, 10, 18, 129, 213, 12, 13, 14, 87, 99, 199         │
-///  └─────────────────────────────────────────────────────────┘
-///   points_map_ids
-///
-/// Mmap-backed immutable geo index.
-///
-/// On-disk state (`counts_per_hash.bin`, `points_map.bin`, `points_map_ids.bin`,
-/// `deleted.bin`, `point_to_values.*`, etc.) is written once during
-/// [`Self::build`] and not mutated afterwards: `deleted.bin` records only the
-/// points whose payload was empty at build time.
-///
-/// Runtime deletions live in the in-memory `Storage::deleted` bitvec. They are
-/// **not persisted** — [`Self::flusher`] is a no-op and [`Self::remove_point`]
-/// only updates the in-memory bitvec. Callers must re-supply the authoritative
-/// deletion set (typically `id_tracker.deleted_point_bitslice()`) via the
-/// `deleted_points` argument to [`Self::open`] on reload.
-pub struct StoredGeoMapIndex<S: UniversalRead> {
-    path: PathBuf,
-    pub(super) storage: Storage<S>,
-    pub(super) deleted_count: usize,
-    points_values_count: usize,
-    max_values_per_point: usize,
-    is_on_disk: bool,
-}
-
-pub(super) struct Storage<S: UniversalRead> {
-    /// Stores GeoHash, points count and values count.
-    /// Sorted by geohash, so we binary search the region.
-    pub(super) counts_per_hash: TypedStorage<S, Counts>,
-    /// Stores GeoHash and associated range of offsets in the points_map_ids.
-    /// Sorted by geohash, so we binary search the region.
-    pub(super) points_map: TypedStorage<S, PointKeyValue>,
-    /// A storage of associations between geo-hashes and point ids. (See the diagram above)
-    pub(super) points_map_ids: TypedStorage<S, PointOffsetType>,
-    /// One-to-many mapping of the PointOffsetType to the GeoPoint.
-    pub(super) point_to_values: StoredPointToValues<GeoPoint, S>,
-    /// In-memory deletion bitmap. Reconstructed at load time as the union of
-    /// the build-time empty-payload bits read from `deleted.bin` and the
-    /// segment-level deleted bitslice supplied by the id-tracker. Not persisted.
-    pub(super) deleted: BitVec,
-}
-
-impl<S: UniversalRead> Storage<S> {
-    pub(crate) fn ram_usage_bytes(&self) -> usize {
-        let Self {
-            counts_per_hash,
-            points_map,
-            points_map_ids,
-            point_to_values,
-            deleted,
-        } = self;
-
-        counts_per_hash.ram_usage_bytes()
-            + points_map.ram_usage_bytes()
-            + points_map_ids.ram_usage_bytes()
-            + point_to_values.ram_usage_bytes()
-            + deleted.capacity().div_ceil(u8::BITS as usize)
-    }
-}
-
-#[derive(Debug, Clone, Serialize, Deserialize)]
-struct StoredGeoMapIndexStat {
-    points_values_count: usize,
-    max_values_per_point: usize,
-}
 
 impl<S: UniversalRead> StoredGeoMapIndex<S> {
     pub fn build(
@@ -347,7 +250,7 @@ impl<S: UniversalRead> StoredGeoMapIndex<S> {
         }
     }
 
-    pub(super) fn points_per_hash(
+    pub(in super::super) fn points_per_hash(
         &self,
         filter: impl Fn(&(GeoHash, usize)) -> bool,
     ) -> OperationResult<Vec<(GeoHash, usize)>> {
@@ -472,7 +375,7 @@ impl<S: UniversalRead> StoredGeoMapIndex<S> {
     }
 
     /// Return all unique point IDs which have any of the geo-hash prefixes.
-    pub(super) fn all_points(
+    pub(in super::super) fn all_points(
         &self,
         mut geo_hashes: Vec<GeoHash>,
     ) -> OperationResult<AHashSet<PointOffsetType>> {
@@ -648,102 +551,5 @@ impl<S: UniversalRead> StoredGeoMapIndex<S> {
 
     pub(crate) fn ram_usage_bytes(&self) -> usize {
         self.storage.ram_usage_bytes()
-    }
-}
-
-impl<S: UniversalRead> GeoMapIndexRead for StoredGeoMapIndex<S> {
-    fn points_count(&self) -> usize {
-        StoredGeoMapIndex::points_count(self)
-    }
-
-    fn points_values_count(&self) -> usize {
-        StoredGeoMapIndex::points_values_count(self)
-    }
-
-    fn max_values_per_point(&self) -> usize {
-        StoredGeoMapIndex::max_values_per_point(self)
-    }
-
-    fn points_of_hash(
-        &self,
-        hash: GeoHash,
-        hw_counter: &HardwareCounterCell,
-    ) -> OperationResult<usize> {
-        StoredGeoMapIndex::points_of_hash(self, hash, hw_counter)
-    }
-
-    fn values_of_hash(
-        &self,
-        hash: GeoHash,
-        hw_counter: &HardwareCounterCell,
-    ) -> OperationResult<usize> {
-        StoredGeoMapIndex::values_of_hash(self, hash, hw_counter)
-    }
-
-    fn check_values_any(
-        &self,
-        idx: PointOffsetType,
-        hw_counter: &HardwareCounterCell,
-        check_fn: &dyn Fn(&GeoPoint) -> bool,
-    ) -> bool {
-        StoredGeoMapIndex::check_values_any(self, idx, hw_counter, |p| check_fn(p))
-    }
-
-    fn values_count(&self, idx: PointOffsetType) -> usize {
-        StoredGeoMapIndex::values_count(self, idx)
-    }
-
-    fn get_values(&self, idx: PointOffsetType) -> Option<Box<dyn Iterator<Item = GeoPoint> + '_>> {
-        StoredGeoMapIndex::get_values(self, idx)
-            .map(|iter| Box::new(iter) as Box<dyn Iterator<Item = GeoPoint> + '_>)
-    }
-
-    fn iterator(
-        &self,
-        values: Vec<GeoHash>,
-    ) -> OperationResult<Box<dyn Iterator<Item = PointOffsetType> + '_>> {
-        let points = self.all_points(values)?;
-        Ok(Box::new(points.into_iter()))
-    }
-
-    fn points_per_hash_filtered(
-        &self,
-        filter: &dyn Fn(&(GeoHash, usize)) -> bool,
-    ) -> OperationResult<Vec<(GeoHash, usize)>> {
-        StoredGeoMapIndex::points_per_hash(self, |pair| filter(pair))
-    }
-
-    fn get_storage_type(&self) -> StorageType {
-        StorageType::Mmap {
-            is_on_disk: self.is_on_disk(),
-        }
-    }
-
-    fn ram_usage_bytes(&self) -> usize {
-        StoredGeoMapIndex::ram_usage_bytes(self)
-    }
-
-    fn is_on_disk(&self) -> bool {
-        StoredGeoMapIndex::is_on_disk(self)
-    }
-
-    fn populate(&self) -> OperationResult<()> {
-        StoredGeoMapIndex::populate(self)
-    }
-
-    fn clear_cache(&self) -> OperationResult<()> {
-        StoredGeoMapIndex::clear_cache(self)
-    }
-
-    fn files(&self) -> Vec<PathBuf> {
-        StoredGeoMapIndex::files(self)
-    }
-
-    fn immutable_files(&self) -> Vec<PathBuf> {
-        StoredGeoMapIndex::immutable_files(self)
-    }
-
-    fn telemetry_index_type(&self) -> &'static str {
-        "mmap_geo"
     }
 }
