--- lib/segment/src/index/field_index/geo_index/mutable_geo_index.rs
@@ -1,692 +1,0 @@
-use std::cmp::max;
-use std::collections::BTreeMap;
-use std::path::PathBuf;
-
-use ahash::AHashSet;
-use common::counter::hardware_counter::HardwareCounterCell;
-use common::types::PointOffsetType;
-use gridstore::Gridstore;
-use gridstore::config::StorageOptions;
-use itertools::Itertools;
-
-use super::read_ops::GeoMapIndexRead;
-use crate::common::Flusher;
-use crate::common::operation_error::{OperationError, OperationResult};
-use crate::index::field_index::geo_hash::{GeoHash, encode_max_precision};
-use crate::index::payload_config::StorageType;
-use crate::types::{GeoPoint, RawGeoPoint};
-
-/// Default options for Gridstore storage
-const GRIDSTORE_OPTIONS: StorageOptions = StorageOptions {
-    // Size of geo point values in index
-    block_size_bytes: Some(size_of::<RawGeoPoint>()),
-    // Compressing geo point values is unreasonable
-    compression: Some(gridstore::config::Compression::None),
-    // Scale page size down with block size, prevents overhead of first page when there's (almost) no values
-    page_size_bytes: Some(size_of::<RawGeoPoint>() * 8192 * 32), // 4 to 8 MiB = block_size * region_blocks * regions,
-    region_size_blocks: None,
-};
-
-pub struct MutableGeoMapIndex {
-    in_memory_index: InMemoryGeoMapIndex,
-    storage: Storage,
-}
-
-enum Storage {
-    Gridstore(Gridstore<Vec<RawGeoPoint>>),
-}
-
-pub struct InMemoryGeoMapIndex {
-    /*
-    {
-        "d": 10,
-        "dr": 10,
-        "dr5": 4,
-        "dr5r": 3,
-        "dr5ru": 1,
-        "dr5rr": 2,
-        ...
-    }
-     */
-    pub points_per_hash: BTreeMap<GeoHash, usize>,
-    pub values_per_hash: BTreeMap<GeoHash, usize>,
-    /*
-    {
-        "dr5ru": {1},
-        "dr5rr": {2, 3},
-        ...
-    }
-     */
-    pub points_map: BTreeMap<GeoHash, AHashSet<PointOffsetType>>,
-    pub point_to_values: Vec<Vec<GeoPoint>>,
-    pub points_count: usize,
-    pub points_values_count: usize,
-    pub max_values_per_point: usize,
-}
-
-impl MutableGeoMapIndex {
-    /// Open and load mutable geo index from Gridstore storage
-    ///
-    /// The `create_if_missing` parameter indicates whether to create a new Gridstore if it does
-    /// not exist. If false and files don't exist, the load function will indicate nothing could be
-    /// loaded.
-    pub fn open_gridstore(path: PathBuf, create_if_missing: bool) -> OperationResult<Option<Self>> {
-        let store = if create_if_missing {
-            Gridstore::open_or_create(path, GRIDSTORE_OPTIONS).map_err(|err| {
-                OperationError::service_error(format!(
-                    "failed to open mutable geo index on gridstore: {err}"
-                ))
-            })?
-        } else if path.exists() {
-            Gridstore::open(path).map_err(|err| {
-                OperationError::service_error(format!(
-                    "failed to open mutable geo index on gridstore: {err}"
-                ))
-            })?
-        } else {
-            // Files don't exist, cannot load
-            return Ok(None);
-        };
-
-        // Load in-memory index from Gridstore
-        let mut in_memory_index = InMemoryGeoMapIndex::new();
-        let hw_counter = HardwareCounterCell::disposable();
-        let hw_counter_ref = hw_counter.ref_payload_index_io_write_counter();
-        store
-            .iter::<_, OperationError>(
-                |idx, values: Vec<RawGeoPoint>| {
-                    let geo_points = values.into_iter().map(GeoPoint::from).collect::<Vec<_>>();
-                    let geo_hashes = geo_points
-                        .iter()
-                        .map(|geo_point| {
-                            encode_max_precision(geo_point.lon.0, geo_point.lat.0).map_err(|e| {
-                                OperationError::service_error(format!("Malformed geo points: {e}"))
-                            })
-                        })
-                        .collect::<Result<Vec<_>, _>>()?;
-
-                    for geo_point in geo_points {
-                        if in_memory_index.point_to_values.len() <= idx as usize {
-                            in_memory_index
-                                .point_to_values
-                                .resize_with(idx as usize + 1, Vec::new);
-                        }
-
-                        if in_memory_index.point_to_values[idx as usize].is_empty() {
-                            in_memory_index.points_count += 1;
-                        }
-
-                        in_memory_index.point_to_values[idx as usize].push(geo_point);
-                        in_memory_index.points_values_count += 1;
-                    }
-
-                    in_memory_index.max_values_per_point =
-                        max(in_memory_index.max_values_per_point, geo_hashes.len());
-                    in_memory_index.increment_hash_point_counts(&geo_hashes);
-                    for geo_hash in geo_hashes {
-                        in_memory_index.increment_hash_value_counts(geo_hash);
-                        in_memory_index
-                            .points_map
-                            .entry(geo_hash)
-                            .or_default()
-                            .insert(idx);
-                    }
-
-                    Ok(true)
-                },
-                hw_counter_ref,
-            )
-            .map_err(|err| {
-                OperationError::service_error(format!(
-                    "Failed to load mutable geo index from gridstore: {err}"
-                ))
-            })?;
-
-        Ok(Some(Self {
-            in_memory_index,
-            storage: Storage::Gridstore(store),
-        }))
-    }
-
-    #[expect(dead_code)] // FIXME(rocksdb): leftover after removing rocksdb
-    #[inline]
-    pub(super) fn clear(&mut self) -> OperationResult<()> {
-        match &mut self.storage {
-            Storage::Gridstore(store) => store.clear().map_err(|err| {
-                OperationError::service_error(format!("Failed to clear mutable geo index: {err}",))
-            }),
-        }
-    }
-
-    #[inline]
-    pub(super) fn wipe(self) -> OperationResult<()> {
-        match self.storage {
-            Storage::Gridstore(store) => store.wipe().map_err(|err| {
-                OperationError::service_error(format!("Failed to wipe mutable geo index: {err}",))
-            }),
-        }
-    }
-
-    /// Clear cache
-    ///
-    /// Only clears cache of Gridstore storage if used. Does not clear in-memory representation of
-    /// index.
-    pub fn clear_cache(&self) -> OperationResult<()> {
-        match &self.storage {
-            Storage::Gridstore(index) => index.clear_cache().map_err(|err| {
-                OperationError::service_error(format!(
-                    "Failed to clear mutable geo index gridstore cache: {err}"
-                ))
-            }),
-        }
-    }
-
-    #[inline]
-    pub(super) fn files(&self) -> Vec<PathBuf> {
-        match &self.storage {
-            Storage::Gridstore(store) => store.files(),
-        }
-    }
-
-    #[inline]
-    pub(super) fn flusher(&self) -> Flusher {
-        match &self.storage {
-            Storage::Gridstore(store) => {
-                let storage_flusher = store.flusher();
-                Box::new(move || storage_flusher().map_err(OperationError::from))
-            }
-        }
-    }
-
-    pub fn add_many_geo_points(
-        &mut self,
-        idx: PointOffsetType,
-        values: &[GeoPoint],
-        hw_counter: &HardwareCounterCell,
-    ) -> OperationResult<()> {
-        // Update persisted storage
-        match &mut self.storage {
-            // We cannot store empty value, then delete instead
-            Storage::Gridstore(store) if values.is_empty() => {
-                store.delete_value(idx)?;
-            }
-            Storage::Gridstore(store) => {
-                let hw_counter_ref = hw_counter.ref_payload_index_io_write_counter();
-                let values = values
-                    .iter()
-                    .cloned()
-                    .map(RawGeoPoint::from)
-                    .collect::<Vec<_>>();
-                store
-                    .put_value(idx, &values, hw_counter_ref)
-                    .map_err(|err| {
-                        OperationError::service_error(format!(
-                            "failed to put value in mutable geo index gridstore: {err}"
-                        ))
-                    })?;
-            }
-        }
-
-        self.in_memory_index
-            .add_many_geo_points(idx, values, hw_counter)
-    }
-
-    pub fn remove_point(&mut self, idx: PointOffsetType) -> OperationResult<()> {
-        // Update persisted storage
-        match &mut self.storage {
-            Storage::Gridstore(store) => {
-                store.delete_value(idx)?;
-            }
-        }
-
-        self.in_memory_index.remove_point(idx)
-    }
-
-    pub fn points_count(&self) -> usize {
-        self.in_memory_index.points_count
-    }
-
-    pub fn points_values_count(&self) -> usize {
-        self.in_memory_index.points_values_count
-    }
-
-    pub fn max_values_per_point(&self) -> usize {
-        self.in_memory_index.max_values_per_point
-    }
-
-    pub fn into_in_memory_index(self) -> InMemoryGeoMapIndex {
-        self.in_memory_index
-    }
-
-    pub fn get_values(&self, idx: u32) -> Option<impl Iterator<Item = &GeoPoint> + '_> {
-        self.in_memory_index
-            .point_to_values
-            .get(idx as usize)
-            .map(|v| v.iter())
-    }
-
-    pub fn check_values_any(
-        &self,
-        idx: PointOffsetType,
-        check_fn: impl Fn(&GeoPoint) -> bool,
-    ) -> bool {
-        self.in_memory_index.check_values_any(idx, check_fn)
-    }
-
-    pub fn values_count(&self, idx: PointOffsetType) -> usize {
-        self.in_memory_index.values_count(idx)
-    }
-
-    pub fn points_per_hash(&self) -> impl Iterator<Item = (GeoHash, usize)> + '_ {
-        self.in_memory_index.points_per_hash()
-    }
-
-    pub fn points_of_hash(&self, hash: GeoHash) -> usize {
-        self.in_memory_index.points_of_hash(hash)
-    }
-
-    pub fn values_of_hash(&self, hash: GeoHash) -> usize {
-        self.in_memory_index.values_of_hash(hash)
-    }
-
-    pub fn stored_sub_regions(&self, geo: GeoHash) -> impl Iterator<Item = PointOffsetType> + '_ {
-        self.in_memory_index.stored_sub_regions(geo)
-    }
-
-    pub fn storage_type(&self) -> StorageType {
-        match &self.storage {
-            Storage::Gridstore(_) => StorageType::Gridstore,
-        }
-    }
-}
-
-impl Default for InMemoryGeoMapIndex {
-    fn default() -> Self {
-        Self::new()
-    }
-}
-
-impl InMemoryGeoMapIndex {
-    pub fn new() -> Self {
-        Self {
-            points_per_hash: Default::default(),
-            values_per_hash: Default::default(),
-            points_map: Default::default(),
-            point_to_values: vec![],
-            points_count: 0,
-            points_values_count: 0,
-            max_values_per_point: 0,
-        }
-    }
-
-    pub fn check_values_any(
-        &self,
-        idx: PointOffsetType,
-        check_fn: impl Fn(&GeoPoint) -> bool,
-    ) -> bool {
-        self.point_to_values
-            .get(idx as usize)
-            .map(|values| values.iter().any(check_fn))
-            .unwrap_or(false)
-    }
-
-    pub fn values_count(&self, idx: PointOffsetType) -> usize {
-        self.point_to_values
-            .get(idx as usize)
-            .map(Vec::len)
-            .unwrap_or_default()
-    }
-
-    pub fn points_per_hash(&self) -> impl Iterator<Item = (GeoHash, usize)> + '_ {
-        self.points_per_hash
-            .iter()
-            .map(|(&hash, &count)| (hash, count))
-    }
-
-    pub fn points_of_hash(&self, hash: GeoHash) -> usize {
-        self.points_per_hash.get(&hash).copied().unwrap_or(0)
-    }
-
-    pub fn values_of_hash(&self, hash: GeoHash) -> usize {
-        self.values_per_hash.get(&hash).copied().unwrap_or(0)
-    }
-
-    pub fn remove_point(&mut self, idx: PointOffsetType) -> OperationResult<()> {
-        if self.point_to_values.len() <= idx as usize {
-            return Ok(()); // Already removed or never actually existed
-        }
-
-        let removed_geo_points = std::mem::take(&mut self.point_to_values[idx as usize]);
-
-        if removed_geo_points.is_empty() {
-            return Ok(());
-        }
-
-        self.points_count -= 1;
-        self.points_values_count -= removed_geo_points.len();
-        let mut removed_geo_hashes = AHashSet::with_capacity(removed_geo_points.len());
-
-        for removed_geo_point in removed_geo_points {
-            let removed_geo_hash: GeoHash =
-                encode_max_precision(removed_geo_point.lon.0, removed_geo_point.lat.0).map_err(
-                    |e| OperationError::service_error(format!("Malformed geo points: {e}")),
-                )?;
-            if !removed_geo_hashes.insert(removed_geo_hash) {
-                continue;
-            }
-
-            let is_last = if let Some(hash_ids) = self.points_map.get_mut(&removed_geo_hash) {
-                hash_ids.remove(&idx);
-                hash_ids.is_empty()
-            } else {
-                debug_assert!(
-                    false,
-                    "Geo index error: no points for hash {removed_geo_hash} was found",
-                );
-                false
-            };
-
-            if is_last {
-                self.points_map.remove(&removed_geo_hash);
-            }
-
-            self.decrement_hash_value_counts(removed_geo_hash);
-        }
-
-        self.decrement_hash_point_counts(removed_geo_hashes);
-        Ok(())
-    }
-
-    pub fn add_many_geo_points(
-        &mut self,
-        idx: PointOffsetType,
-        values: &[GeoPoint],
-        hw_counter: &HardwareCounterCell,
-    ) -> OperationResult<()> {
-        if values.is_empty() {
-            return Ok(());
-        }
-
-        if self.point_to_values.len() <= idx as usize {
-            // That's a smart reallocation
-            self.point_to_values.resize_with(idx as usize + 1, Vec::new);
-        }
-
-        self.point_to_values[idx as usize] = values.to_vec();
-
-        let mut geo_hashes = vec![];
-
-        let mut hw_cell_wb = hw_counter
-            .payload_index_io_write_counter()
-            .write_back_counter();
-
-        for added_point in values {
-            let added_geo_hash: GeoHash =
-                encode_max_precision(added_point.lon.0, added_point.lat.0).map_err(|e| {
-                    OperationError::service_error(format!("Malformed geo points: {e}"))
-                })?;
-
-            hw_cell_wb.incr_delta(size_of_val(&added_geo_hash));
-
-            geo_hashes.push(added_geo_hash);
-        }
-
-        for &geo_hash in &geo_hashes {
-            self.points_map.entry(geo_hash).or_default().insert(idx);
-
-            self.increment_hash_value_counts(geo_hash);
-        }
-
-        hw_cell_wb.incr_delta(geo_hashes.len() * size_of::<PointOffsetType>());
-
-        self.increment_hash_point_counts(&geo_hashes);
-
-        self.points_values_count += values.len();
-        self.points_count += 1;
-        self.max_values_per_point = self.max_values_per_point.max(values.len());
-        Ok(())
-    }
-
-    /// Returns an iterator over all point IDs which have the `geohash` prefix.
-    /// Note. Point ID may be repeated multiple times in the iterator.
-    pub fn stored_sub_regions(&self, geo: GeoHash) -> impl Iterator<Item = PointOffsetType> + '_ {
-        self.points_map
-            .range(geo..)
-            .take_while(move |(p, _h)| p.starts_with(geo))
-            .flat_map(|(_, points)| points.iter().copied())
-    }
-
-    fn increment_hash_value_counts(&mut self, geo_hash: GeoHash) {
-        for i in 0..=geo_hash.len() {
-            let sub_geo_hash = geo_hash.truncate(i);
-            match self.values_per_hash.get_mut(&sub_geo_hash) {
-                None => {
-                    self.values_per_hash.insert(sub_geo_hash, 1);
-                }
-                Some(count) => {
-                    *count += 1;
-                }
-            };
-        }
-    }
-
-    fn increment_hash_point_counts(&mut self, geo_hashes: &[GeoHash]) {
-        let mut seen_hashes: AHashSet<GeoHash> = Default::default();
-
-        for geo_hash in geo_hashes {
-            for i in 0..=geo_hash.len() {
-                let sub_geo_hash = geo_hash.truncate(i);
-                if seen_hashes.contains(&sub_geo_hash) {
-                    continue;
-                }
-                seen_hashes.insert(sub_geo_hash);
-                match self.points_per_hash.get_mut(&sub_geo_hash) {
-                    None => {
-                        self.points_per_hash.insert(sub_geo_hash, 1);
-                    }
-                    Some(count) => {
-                        *count += 1;
-                    }
-                };
-            }
-        }
-    }
-
-    fn decrement_hash_value_counts(&mut self, geo_hash: GeoHash) {
-        for i in 0..=geo_hash.len() {
-            let sub_geo_hash = geo_hash.truncate(i);
-            match self.values_per_hash.get_mut(&sub_geo_hash) {
-                None => {
-                    debug_assert!(
-                        false,
-                        "Hash value count is not found for hash: {sub_geo_hash}",
-                    );
-                    self.values_per_hash.insert(sub_geo_hash, 0);
-                }
-                Some(count) => {
-                    *count -= 1;
-                }
-            };
-        }
-    }
-
-    fn decrement_hash_point_counts(&mut self, geo_hashes: impl IntoIterator<Item = GeoHash>) {
-        let mut seen_hashes: AHashSet<GeoHash> = Default::default();
-        for geo_hash in geo_hashes {
-            for i in 0..=geo_hash.len() {
-                let sub_geo_hash = geo_hash.truncate(i);
-                if seen_hashes.contains(&sub_geo_hash) {
-                    continue;
-                }
-                seen_hashes.insert(sub_geo_hash);
-                match self.points_per_hash.get_mut(&sub_geo_hash) {
-                    None => {
-                        debug_assert!(
-                            false,
-                            "Hash point count is not found for hash: {sub_geo_hash}",
-                        );
-                        self.points_per_hash.insert(sub_geo_hash, 0);
-                    }
-                    Some(count) => {
-                        *count -= 1;
-                    }
-                };
-            }
-        }
-    }
-}
-
-impl MutableGeoMapIndex {
-    /// Approximate RAM usage in bytes for in-memory index structures.
-    pub fn ram_usage_bytes(&self) -> usize {
-        let Self {
-            in_memory_index,
-            storage: _,
-        } = self;
-        in_memory_index.ram_usage_bytes()
-    }
-}
-
-impl GeoMapIndexRead for MutableGeoMapIndex {
-    fn points_count(&self) -> usize {
-        MutableGeoMapIndex::points_count(self)
-    }
-
-    fn points_values_count(&self) -> usize {
-        MutableGeoMapIndex::points_values_count(self)
-    }
-
-    fn max_values_per_point(&self) -> usize {
-        MutableGeoMapIndex::max_values_per_point(self)
-    }
-
-    fn points_of_hash(
-        &self,
-        hash: GeoHash,
-        _hw_counter: &HardwareCounterCell,
-    ) -> OperationResult<usize> {
-        Ok(MutableGeoMapIndex::points_of_hash(self, hash))
-    }
-
-    fn values_of_hash(
-        &self,
-        hash: GeoHash,
-        _hw_counter: &HardwareCounterCell,
-    ) -> OperationResult<usize> {
-        Ok(MutableGeoMapIndex::values_of_hash(self, hash))
-    }
-
-    fn check_values_any(
-        &self,
-        idx: PointOffsetType,
-        _hw_counter: &HardwareCounterCell,
-        check_fn: &dyn Fn(&GeoPoint) -> bool,
-    ) -> bool {
-        MutableGeoMapIndex::check_values_any(self, idx, |p| check_fn(p))
-    }
-
-    fn values_count(&self, idx: PointOffsetType) -> usize {
-        MutableGeoMapIndex::values_count(self, idx)
-    }
-
-    fn get_values(&self, idx: PointOffsetType) -> Option<Box<dyn Iterator<Item = GeoPoint> + '_>> {
-        MutableGeoMapIndex::get_values(self, idx)
-            .map(|iter| Box::new(iter.copied()) as Box<dyn Iterator<Item = GeoPoint> + '_>)
-    }
-
-    fn iterator(
-        &self,
-        values: Vec<GeoHash>,
-    ) -> OperationResult<Box<dyn Iterator<Item = PointOffsetType> + '_>> {
-        Ok(Box::new(
-            values
-                .into_iter()
-                .flat_map(|top_geo_hash| self.stored_sub_regions(top_geo_hash))
-                .unique(),
-        ))
-    }
-
-    fn points_per_hash_filtered(
-        &self,
-        filter: &dyn Fn(&(GeoHash, usize)) -> bool,
-    ) -> OperationResult<Vec<(GeoHash, usize)>> {
-        Ok(MutableGeoMapIndex::points_per_hash(self)
-            .filter(|pair| filter(pair))
-            .collect())
-    }
-
-    fn get_storage_type(&self) -> StorageType {
-        MutableGeoMapIndex::storage_type(self)
-    }
-
-    fn ram_usage_bytes(&self) -> usize {
-        MutableGeoMapIndex::ram_usage_bytes(self)
-    }
-
-    fn is_on_disk(&self) -> bool {
-        false
-    }
-
-    fn populate(&self) -> OperationResult<()> {
-        Ok(())
-    }
-
-    fn clear_cache(&self) -> OperationResult<()> {
-        MutableGeoMapIndex::clear_cache(self)
-    }
-
-    fn files(&self) -> Vec<PathBuf> {
-        MutableGeoMapIndex::files(self)
-    }
-
-    fn immutable_files(&self) -> Vec<PathBuf> {
-        vec![]
-    }
-
-    fn telemetry_index_type(&self) -> &'static str {
-        "mutable_geo"
-    }
-}
-
-impl InMemoryGeoMapIndex {
-    /// Approximate RAM usage in bytes.
-    pub fn ram_usage_bytes(&self) -> usize {
-        let Self {
-            points_per_hash,
-            values_per_hash,
-            points_map,
-            point_to_values,
-            points_count: _,
-            points_values_count: _,
-            max_values_per_point: _,
-        } = self;
-
-        let btree_entry_overhead = std::mem::size_of::<usize>() * 3;
-        let pph_bytes = points_per_hash.len()
-            * (std::mem::size_of::<GeoHash>()
-                + std::mem::size_of::<usize>()
-                + btree_entry_overhead);
-        let vph_bytes = values_per_hash.len()
-            * (std::mem::size_of::<GeoHash>()
-                + std::mem::size_of::<usize>()
-                + btree_entry_overhead);
-        // points_map: BTreeMap entries + AHashSet per entry
-        let hashset_entry_overhead = std::mem::size_of::<u64>() + std::mem::size_of::<usize>();
-        let pm_bytes: usize = points_map
-            .values()
-            .map(|set| {
-                std::mem::size_of::<GeoHash>()
-                    + std::mem::size_of::<AHashSet<PointOffsetType>>()
-                    + btree_entry_overhead
-                    + set.capacity()
-                        * (std::mem::size_of::<PointOffsetType>() + hashset_entry_overhead)
-            })
-            .sum();
-        let ptv_bytes: usize = point_to_values.capacity() * std::mem::size_of::<Vec<GeoPoint>>()
-            + point_to_values
-                .iter()
-                .map(|v| v.capacity() * std::mem::size_of::<GeoPoint>())
-                .sum::<usize>();
-        pph_bytes + vph_bytes + pm_bytes + ptv_bytes
-    }
-}
